refctor: make the LoadingOverlay more reusable
