+++ components/common/LoadingOverlay/index.tsx
@@ -1,28 +1,15 @@
+import React, { FC } from 'react';
 import { motion } from 'framer-motion';
-
-import { WalkthroughMarketState } from '@/types/walkthrough';
 import { fadeIn } from '@/utils/animations';
-import { useWalkthroughContext } from '@/context/WalkthroughContext';
 
-const getOverlayText = (marketState: WalkthroughMarketState) => {
-  if (marketState === WalkthroughMarketState.calculating_winners) {
-    return 'Determining Market Winners';
-  }
-
-  if (marketState === WalkthroughMarketState.distributing_surpluss) {
-    return 'Distributing Market Surplus';
-  }
-
-  if (marketState === WalkthroughMarketState.calculating_final_payments) {
-    return 'Calculating Final Payments';
-  }
+type LoadingOverlayProps = {
+  text?: string;
 };
 
-const LoadingOverlay = () => {
-  const { marketState } = useWalkthroughContext();
-  const overlayText = getOverlayText(marketState);
-
-  if (!overlayText) {
+export const LoadingOverlay: FC<LoadingOverlayProps> = ({
+  text,
+}: LoadingOverlayProps) => {
+  if (!text) {
     return null;
   }
 
@@ -34,9 +21,7 @@ const LoadingOverlay = () => {
       className="absolute top-0 left-0 w-full h-full bg-black/70 flex justify-center items-center"
     >
       <div className="flex flex-col items-center">
-        <p className="text-center text-white text-2xl font-bold">
-          {overlayText}
-        </p>
+        <p className="text-center text-white text-2xl font-bold">{text}</p>
 
         <div className="lds-ellipsis inline-block relative w-[80px] h-[80px]">
           <div className="bg-green-light absolute top-[33px] left-[8px] w-[15px] h-[15px] rounded-full" />
@@ -48,5 +33,3 @@ const LoadingOverlay = () => {
     </motion.div>
   );
 };
-
-export default LoadingOverlay;
