+++ jest.config.js
@@ -6,4 +6,13 @@ module.exports = {
   transform: {
     '^.+\\.(t|j)sx?$': ['@swc/jest'],
   },
+  moduleNameMapper: {
+    '^@/components/(.*)$': '<rootDir>/components/$1',
+    '^@/lib/(.*)$': '<rootDir>/lib/$1',
+    '^@/hooks/(.*)$': '<rootDir>/hooks/$1',
+    '^@/utils/(.*)$': '<rootDir>/utils/$1',
+    '^@/styles/(.*)$': '<rootDir>/styles/$1',
+    '^@/context/(.*)$': '<rootDir>/context/$1',
+    '^@/types/(.*)$': '<rootDir>/types/$1',
+  },
 };
