+++ jest.setup.after-env.js
@@ -1,1 +1,2 @@
 import '@testing-library/jest-dom';
+import 'intersection-observer';
