+++ package.json
@@ -59,6 +59,7 @@
     "eslint-plugin-react": "^7.30.1",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-testing-library": "^5.9.1",
+    "intersection-observer": "^0.12.2",
     "jest": "^29.3.1",
     "jest-environment-jsdom": "^29.3.1",
     "next-router-mock": "^0.8.0",
